requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
altair==4.2.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.1.0
audio-recorder-streamlit==0.0.8
audioread==3.0.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmaes==0.10.0
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.1
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.5.0
docopt==0.6.2
einops==0.6.1
entrypoints==0.4
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.17.2
HyperPyYAML==1.2.2
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.1
idna==3.4
importlib-metadata==6.8.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
llvmlite==0.41.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
networkx==3.1
numba==0.58.0
numpy==1.25.2
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
optuna==3.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
pooch==1.7.0
primePy==1.3
protobuf==3.20.3
psutil==5.9.5
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio@d169fd37708d5f913b71841f84e93dd8cbefd174
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==13.0.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-learn==1.3.1
scipy==1.11.2
semver==3.0.1
sentencepiece==0.1.99
shellingham==1.5.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
speechbrain==0.5.15
SQLAlchemy==2.0.21
starlette==0.27.0
starsessions==1.3.0
streamlit==1.17.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==1.2.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.22.0
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
